LayoutTests/http/tests/security/xssAuditor/xss-protection-parsing-01-expected.txt

CONSOLE MESSAGE: line 1: Refused to execute a JavaScript script. Source code of script found within request.

This tests our parsing of the X-XSS-Protection header.



--------
Frame: 'frame'
--------
